refactor(server): split off hospital related code into its own file
